Scrolling smoother, Screenshot updated

+++ MAVGCL/src/com/comino/flight/widgets/charts/control/IChartControl.java
@@ -17,13 +17,14 @@
 package com.comino.flight.widgets.charts.control;
 
 import javafx.beans.property.BooleanProperty;
+import javafx.beans.property.DoubleProperty;
 import javafx.beans.property.IntegerProperty;
 
 public interface IChartControl {
 
 	public BooleanProperty getCollectingProperty();
 	public IntegerProperty getTimeFrameProperty();
-	public IntegerProperty getScrollProperty();
+	public DoubleProperty getScrollProperty();
 
 
 }
